+++ assets/js/blocks/product-collection/utils.tsx
@@ -2,9 +2,6 @@
  * External dependencies
  */
 import { BlockEditProps } from '@wordpress/blocks';
-import { useSelect } from '@wordpress/data';
-import { store as coreStore } from '@wordpress/core-data';
-import { Taxonomy } from '@wordpress/core-data/src/entity-types';
 
 /**
  * Internal dependencies
@@ -29,21 +26,3 @@ export function setQueryAttribute(
 		},
 	} );
 }
-
-/**
- * Hook that returns the taxonomies associated with product post type.
- *
- * @return {Object[]} An array of the taxonomies.
- */
-export const useTaxonomies = (): Taxonomy[] => {
-	const taxonomies = useSelect( ( select ) => {
-		const { getTaxonomies } = select( coreStore );
-		const filteredTaxonomies: Taxonomy[] = getTaxonomies( {
-			type: 'product',
-			per_page: -1,
-			context: 'view',
-		} );
-		return filteredTaxonomies;
-	}, [] );
-	return taxonomies;
-};
